Added pipeline scripts
All scripts use Google Cloud Platform and PySpark. I used Google Dataproc to create these scripts.

ExtractLoad script extracts the data and loads it into a bucket
Clean script cleans the data as best as it can before dimensional modeling
DimensionCreation script loads all the clean data and makes all the dimensions and facts tables. It then saves them to their own folders in the bucket

ISSUE:
Currently, dimension creation is still unfinished. To finish it, the facts_gdp table needs to connect to dim_census_location. Currently, there is a rough setup to do this by joining both the facts_gdp and dim_census_location through common counties for both tables. However, it creates an absurdly large amount of rows in the process. To understand why, here is an example.

For example, if there are 51 rows that contain the county "MARION" in facts_gdp, and there are 57,990 rows that contain it in dim_census_location, then a total of 2,957,490 rows will be created. There are no nulls. If left as is, then joining the 2 tables will lead to up to 70 billion records created when accounting for all counties. Obviously, this is ridiculously large and likely should be changed.